python/monitor/requirements.txt
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
autopep8==1.3.3
base58==0.2.5
bitcoin==1.1.42
blessings==1.6.1
bpython==0.17
bumpversion==0.5.3
byteplay3==3.5.0
certifi==2017.11.5
cffi==1.11.4
chardet==3.0.4
colorlog==3.1.0
constantly==15.1.0
coverage==4.4.2
coveralls==1.2.0
cryptography==2.1.4
curtsies==0.2.11
cycler==0.10.0
docopt==0.6.2
ecdsa==0.13
Events==0.3
gevent==1.2.2
greenlet==0.4.12
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
klein==17.10.0
logzero==1.3.1
memory-profiler==0.51.0
mmh3==2.5.1
mock==2.0.0
mpmath==1.0.0
neo-boa==0.2.2
neo-python-rpc==0.1.8
neocore==0.3.5
numpy==1.14.0
pbr==3.1.1
peewee==2.10.2
pluggy==0.6.0
plyvel==1.0.4
prompt-toolkit==1.0.15
psutil==5.4.3
py==1.5.2
pycodestyle==2.3.1
pycparser==2.18
pycrypto==2.6.1
Pygments==2.2.0
pymitter==0.2.3
Pympler==0.5
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.3
requests==2.18.4
scrypt==0.8.0
six==1.11.0
tqdm==4.19.5
Twisted==17.9.0
urllib3==1.22
virtualenv==15.1.0
wcwidth==0.1.7
Werkzeug==0.14.1
zope.interface==4.4.3
google-cloud-pubsub==0.31.0
